refactor: Use ENDPOINT instead of SPARQL_ENDPOINT (env)
