+++ app/.env.development
@@ -1,5 +1,5 @@
 DATABASE_URL=postgres://postgres:password@localhost:5432/visualization_tool
+ENDPOINT=sparql+https://lindas.admin.ch/query
 SPARQL_GEO_ENDPOINT=https://geo.ld.admin.ch/query
-SPARQL_ENDPOINT=https://int.lindas.admin.ch/query
 SPARQL_EDITOR=https://int.lindas.admin.ch/sparql/
 GRAPHQL_ENDPOINT=/api/graphql
