+++ app/domain/env.ts
@@ -20,10 +20,10 @@ export const PUBLIC_URL = (
   ""
 ).replace(/\/$/, "");
 
-export const SPARQL_ENDPOINT =
-  clientEnv?.SPARQL_ENDPOINT ??
-  process.env.SPARQL_ENDPOINT ??
-  "https://int.lindas.admin.ch/query";
+export const ENDPOINT =
+  clientEnv?.ENDPOINT ??
+  process.env.ENDPOINT ??
+  "sparql+https://lindas.admin.ch/query";
 
 export const SPARQL_GEO_ENDPOINT =
   clientEnv?.SPARQL_GEO_ENDPOINT ??
