+++ app/rdf/sparql-client.ts
@@ -1,10 +1,14 @@
 import { DatasetCore, Quad, Stream } from "rdf-js";
 import StreamClient from "sparql-http-client";
 
-import { SPARQL_ENDPOINT } from "../domain/env";
+import { convertEndpointToSource } from "@/graphql/resolvers/utils";
+
+import { ENDPOINT } from "../domain/env";
+
+export const DEFAULT_DATA_SOURCE = convertEndpointToSource(ENDPOINT);
 
 export const sparqlClientStream = new StreamClient({
-  endpointUrl: SPARQL_ENDPOINT,
+  endpointUrl: DEFAULT_DATA_SOURCE.url,
 });
 
 export const fromStream = (
